#23 - GatewayAPI and Frontend Refactoring for Improved Scalability and Performance
