--- GatewayAPI/Dockerfile
@@ -1,14 +1,0 @@
-# syntax=docker/dockerfile:1
-
-FROM python:3.8-slim-buster
-
-WORKDIR /app
-
-COPY requirements.txt requirements.txt
-RUN pip3 install -r requirements.txt
-ENV FLASK_APP="gateway_api.py"
-ENV DEBUG="True"
-
-COPY . .
-
-CMD ["python3", "-m" , "flask", "run", "--host=0.0.0.0"]
