--- GatewayAPI/requirements.txt
@@ -1,7 +1,0 @@
-click==8.1.3
-Flask==2.3.1
-itsdangerous==2.1.2
-Jinja2==3.1.2
-MarkupSafe==2.1.2
-Werkzeug==2.3.0
-
